requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi>=0.110.0
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.7
importlib-metadata==7.0.1
iniconfig==2.0.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.1.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.17.2
pytest==8.0.0
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.32.1
regex==2023.12.25
reportlab==4.0.9
requests==2.32.2
psycopg2-binary==2.9.9
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
starlette>=0.40.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado>=6.4.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3>=2.2.2
uvicorn>=0.27.0.post1
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
XlsxWriter==3.1.9
zipp>=3.19.1
icalendar
pydantic-settings==2.2.1